IoT/scr/requirements.txt
Adafruit-Blinka==5.5.4
adafruit-circuitpython-busdevice==5.0.1
adafruit-circuitpython-mlx90640==1.2.0
adafruit-circuitpython-register==1.9.0
Adafruit-PlatformDetect==2.18.2
Adafruit-PureIO==1.1.7
apturl==0.5.2
asn1crypto==0.24.0
beautifulsoup4==4.6.0
blinker==1.4
Brlapi==0.6.6
certifi==2018.1.18
chardet==3.0.4
click==7.1.2
cryptography==2.1.4
cupshelpers==1.0
Cython==0.29.21
decorator==4.1.2
defer==1.0.6
distro==1.5.0
distro-info===0.18ubuntu0.18.04.1
dlib==19.21.0
feedparser==5.2.1
Flask==1.1.2
Flask-Cors==3.0.9
Flask-SocketIO==4.3.1
graphsurgeon==0.4.5
html5lib==0.999999999
httplib2==0.9.2
idna==2.6
itsdangerous==1.1.0
jetson-stats==3.0.1
Jetson.GPIO==2.0.11
Jinja2==2.11.2
joblib==0.17.0
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
louis==3.5.0
lxml==4.2.1
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.1.1
numpy==1.16.4
oauth==1.0.1
oauthlib==2.0.6
onboard==1.4.1
packaging==20.4
pandas==0.22.0
Pillow==8.0.0
protobuf==3.13.0
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
pyftdi==0.52.0
pygobject==3.26.1
PyJWT==1.5.3
pymacaroons==0.13.0
pymongo==3.11.0
PyNaCl==1.1.2
pyparsing==2.4.7
pyRFC3339==1.0
pyserial==3.4
python-apt==1.6.5+ubuntu0.3
python-dateutil==2.6.1
python-debian==0.1.32
python-engineio==3.13.2
python-socketio==4.6.0
pytz==2018.3
pyusb==1.1.0
pyxdg==0.25
PyYAML==3.12
requests==2.18.4
requests-unixsocket==0.1.5
scikit-build==0.11.1
scikit-learn==0.21.3
scipy==1.5.3
SecretStorage==2.3.1
simplejson==3.13.2
six==1.15.0
ssh-import-id==5.7
system-service==0.3
systemd-python==234
tensorrt==7.1.3.0
threadpoolctl==2.1.0
ubuntu-drivers-common==0.0.0
uff==0.6.9
unity-scope-calculator==0.1
unity-scope-chromiumbookmarks==0.1
unity-scope-colourlovers==0.1
unity-scope-devhelp==0.1
unity-scope-firefoxbookmarks==0.1
unity-scope-manpages==0.1
unity-scope-openclipart==0.1
unity-scope-texdoc==0.1
unity-scope-tomboy==0.1
unity-scope-virtualbox==0.1
unity-scope-yelp==0.1
unity-scope-zotero==0.1
urllib3==1.22
urwid==2.0.1
wadllib==1.3.2
webencodings==0.5
Werkzeug==1.0.1
xkit==0.0.0
zope.interface==4.3.2
